Fix Id & IdManager invariant not being maintained in all cases
IDs were not properly reposited on move assignment. Using request() would leave the manager in an invalid state, if the requested ID is larger than the current next value. This commit also simplifies the internal free-list cleanup and adds an internal isValidId function that can be used in unit tests.